+++ products/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
-from core.views import TopicView
+from products.views import TopicView
+
 
 urlpatterns = [
     path('/topic', TopicView.as_view()),
